Update: Use Vector2D

+++ src/Utils.java
@@ -8,7 +8,7 @@ public class Utils {
      * @param max
      * @return
      */
-    public static int clamp(int value, int min, int max) {
+    public static float clamp(float value, float min, float max) {
         if (value < min) {
             return min;
         }
